requirements.txt
Flask>=3.1.2
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
